CCS-65 refactor: Redis 저장 형식을 "유저 식별자 : Refresh Token" 형태로 변경 및 AES 암호화 로직 추가
