+++ backend/src/main/java/com/trend_now/backend/member/application/MemberRedisService.java
@@ -14,16 +14,17 @@ public class MemberRedisService {
 
     private final RedisTemplate<String, String> redisTemplate;
 
+    private static final String REFRESH_TOKEN_PREFIX = "RefreshToken_";
 
     /**
      * Redis에 Refresh Token 저장
-     * - key가 Refresh Token, value가 Member Id가 된다.
-     * - Access Token 재발급 시, key로서 Refresh Token을 받기 위함
+     * - key가 prefix + Member Id, value가 Refresh Token이 된다.
+     * - ex. RefreshToken_member1 : RefreshToken1
      */
-    public void saveRefreshToken(String refreshToken, Long memberId, int refreshTokenExpiration) {
+    public void saveRefreshToken(Long memberId, String refreshToken, int refreshTokenExpiration) {
         redisTemplate.opsForValue().set(
+                REFRESH_TOKEN_PREFIX + memberId.toString(),
                 refreshToken,
-                String.valueOf(memberId),
                 Duration.ofMinutes(refreshTokenExpiration)
         );
 
@@ -32,12 +33,9 @@ public class MemberRedisService {
     }
 
     /**
-     * Redis에 Refresh Token을 이용해 Member Id 조회
+     * Redis에 Member Id를 이용해 Refresh Token 조회
      */
-    public String findMemberIdByRefreshToken(String refreshToken) {
-        String memberId = redisTemplate.opsForValue().get(refreshToken);
-
-        log.info("[RedisUtil.findMemberIdByRefreshToken] memberId={}", memberId);
-        return memberId;
+    public boolean isMemberIdInRedis(String memberId) {
+        return redisTemplate.opsForValue().get(REFRESH_TOKEN_PREFIX + memberId) != null;
     }
 }
